+++ app/src/main/java/com/boardgamegeek/ui/GeekListItemActivity.java
@@ -3,17 +3,20 @@ package com.boardgamegeek.ui;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.support.v7.app.ActionBar;
-import android.text.TextUtils;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
 import android.view.MenuItem;
 
 import com.boardgamegeek.R;
 import com.boardgamegeek.provider.BggContract;
 import com.boardgamegeek.util.ActivityUtils;
+import com.boardgamegeek.util.ImageUtils;
+import com.boardgamegeek.util.ScrimUtils;
+import com.boardgamegeek.util.UIUtils;
 import com.crashlytics.android.answers.Answers;
 import com.crashlytics.android.answers.ContentViewEvent;
 
-public class GeekListItemActivity extends SimpleSinglePaneActivity {
+public class GeekListItemActivity extends HeroTabActivity {
 	private int geekListId;
 	private String geekListTitle;
 	private int objectId;
@@ -32,13 +35,9 @@ public class GeekListItemActivity extends SimpleSinglePaneActivity {
 		objectName = intent.getStringExtra(ActivityUtils.KEY_NAME);
 		url = intent.getStringExtra(ActivityUtils.KEY_OBJECT_URL);
 		isBoardGame = intent.getBooleanExtra(ActivityUtils.KEY_IS_BOARD_GAME, false);
+		int imageId = intent.getIntExtra(ActivityUtils.KEY_IMAGE_ID, BggContract.INVALID_ID);
 
-		if (!TextUtils.isEmpty(geekListTitle)) {
-			final ActionBar actionBar = getSupportActionBar();
-			if (actionBar != null) {
-				actionBar.setTitle(geekListTitle);
-			}
-		}
+		safelySetTitle(objectName);
 
 		if (savedInstanceState == null) {
 			Answers.getInstance().logContentView(new ContentViewEvent()
@@ -46,11 +45,9 @@ public class GeekListItemActivity extends SimpleSinglePaneActivity {
 				.putContentId(String.valueOf(objectId))
 				.putContentName(objectName));
 		}
-	}
 
-	@Override
-	protected Fragment onCreatePane(Intent intent) {
-		return new GeekListItemFragment();
+		ScrimUtils.applyInvertedScrim(scrimView);
+		ImageUtils.safelyLoadImage(toolbarImage, imageId, null);
 	}
 
 	@Override
@@ -82,5 +79,46 @@ public class GeekListItemActivity extends SimpleSinglePaneActivity {
 				return true;
 		}
 		return super.onOptionsItemSelected(item);
+	}
+
+	@Override
+	protected void setUpViewPager() {
+		GeekListItemPagerAdapter adapter = new GeekListItemPagerAdapter(getSupportFragmentManager());
+		viewPager.setAdapter(adapter);
+	}
+
+	private final class GeekListItemPagerAdapter extends FragmentPagerAdapter {
+		public GeekListItemPagerAdapter(FragmentManager fragmentManager) {
+			super(fragmentManager);
+		}
+
+		@Override
+		public CharSequence getPageTitle(int position) {
+			if (position == 0) return getString(R.string.title_description);
+			if (position == 1) return getString(R.string.title_comments);
+			return "";
+		}
+
+		@Override
+		public Fragment getItem(int position) {
+			if (position == 0) {
+				return Fragment.instantiate(
+					GeekListItemActivity.this,
+					GeekListItemFragment.class.getName(),
+					UIUtils.intentToFragmentArguments(getIntent()));
+			}
+			if (position == 1) {
+				return Fragment.instantiate(
+					GeekListItemActivity.this,
+					GeekListCommentsFragment.class.getName(),
+					UIUtils.intentToFragmentArguments(getIntent()));
+			}
+			return null;
+		}
+
+		@Override
+		public int getCount() {
+			return 2;
+		}
 	}
 }